@@ -1,4 +1,4 @@
-class Collaboration < ActiveRecord::Base
+class Collaboration < ApplicationRecord
   belongs_to :user
   belongs_to :year
 end
